+++ src/main/java/io/github/devatherock/artifactory/util/BadgeGenerator.java
@@ -2,20 +2,11 @@ package io.github.devatherock.artifactory.util;
 
 import io.github.devatherock.artifactory.config.ShieldsIOProperties;
 import io.github.devatherock.artifactory.service.ShieldsIOClient;
-import io.micronaut.core.io.IOUtils;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.http.HttpResponse;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpGet;
 
 import javax.inject.Singleton;
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.URI;
-import java.net.URL;
 
 @Slf4j
 @Singleton
@@ -24,7 +15,6 @@ public class BadgeGenerator {
     private static final int WIDTH_ONE_CHAR = 8;
 
     private final ShieldsIOClient shieldsIOClient;
-    private final HttpClient httpClient;
     private final ShieldsIOProperties config;
 
     @SneakyThrows
@@ -34,22 +24,7 @@ public class BadgeGenerator {
 
         if (config.isEnabled()) {
             String shieldIoImageName = String.format("%s-%s-blue.svg", label, value);
-
-            if (config.isCustomHttp()) {
-                // For URL encoding spaces in the path. URLEncoder.encode encodes them into + instead of %20
-                URL url = new URL(config.getUrl() + "/" + shieldIoImageName);
-                URI uri = new URI(url.getProtocol(), url.getUserInfo(), url.getHost(), url.getPort(), url.getPath(),
-                        url.getQuery(), url.getRef());
-                HttpGet httpGet = new HttpGet(uri);
-                HttpResponse httpResponse = httpClient.execute(httpGet);
-
-                try (InputStream stream = httpResponse.getEntity().getContent();
-                     BufferedReader reader = new BufferedReader(new InputStreamReader(stream))) {
-                    badge = IOUtils.readText(reader);
-                }
-            } else {
-                badge = shieldsIOClient.downloadBadge(shieldIoImageName);
-            }
+            badge = shieldsIOClient.downloadBadge(shieldIoImageName);
         } else {
             badge = generateSvg(label, value);
         }